+++ src/components/Vote.tsx
@@ -20,6 +20,7 @@ const Vote = ({
   likes,
   score,
   size = 6,
+  archived = false,
   hideScore = false,
   postindex = undefined,
   postMode = false,
@@ -35,7 +36,7 @@ const Vote = ({
   const castVote = async (e, v) => {
     e?.preventDefault();
     e?.stopPropagation();
-    if (session) {
+    if (session && !archived) {
       let postv;
       if (v === liked) {
         postv = 0;
@@ -56,7 +57,7 @@ const Vote = ({
         setLiked(undefined);
         setVoteScore(calculateScore(score));
       }
-    } else {
+    } else if (!session) {
       context.setLoginModal(true);
     }
   };
@@ -91,11 +92,11 @@ const Vote = ({
       viewBox="0 0 24 24"
       className={
         (liked ? " text-upvote " : "") +
-        ` flex-none cursor-pointer w-${size} h-${size} hover:text-upvote scale-110 hover:scale-100`
+        ` flex-none  w-${size} h-${size} ${!archived ? " cursor-pointer hover:text-upvote scale-110 hover:scale-100" : ""}`
       }
       xmlns="http://www.w3.org/2000/svg"
     >
-      <title>{`unvote ${postMode ? "(a)" : ""}`}</title>
+      <title>{`${!archived ? `unvote ${postMode ? "(a)" : ""}` : "archived"}`}</title>
       <path d="M12.781,2.375C12.4,1.9,11.6,1.9,11.219,2.375l-8,10c-0.24,0.301-0.286,0.712-0.12,1.059C3.266,13.779,3.615,14,4,14h2h2 v3v4c0,0.553,0.447,1,1,1h6c0.553,0,1-0.447,1-1v-5v-2h2h2c0.385,0,0.734-0.221,0.901-0.566c0.166-0.347,0.12-0.758-0.12-1.059 L12.781,2.375z"></path>
     </svg>
   );
@@ -109,11 +110,11 @@ const Vote = ({
       viewBox="0 0 24 24"
       className={
         (liked === false ? " text-downvote " : " ") +
-        ` flex-none cursor-pointer w-${size} h-${size} hover:text-downvote scale-110 hover:scale-100`
+        ` flex-none w-${size} h-${size} ${!archived ? " cursor-pointer } hover:text-downvote scale-110 hover:scale-100" : ""} `
       }
       xmlns="http://www.w3.org/2000/svg"
     >
-      <title>{`unvote ${postMode ? "(v)" : ""}`}</title>
+      <title>{`${!archived ? `unvote ${postMode ? "(v)" : ""}` : "archived"}`}</title>
       <path d="M20.901,10.566C20.734,10.221,20.385,10,20,10h-2h-2V7V3c0-0.553-0.447-1-1-1H9C8.447,2,8,2.447,8,3v5v2H6H4 c-0.385,0-0.734,0.221-0.901,0.566c-0.166,0.347-0.12,0.758,0.12,1.059l8,10C11.409,21.862,11.696,22,12,22 s0.591-0.138,0.781-0.375l8-10C21.021,11.324,21.067,10.913,20.901,10.566z"></path>
     </svg>
   );
@@ -124,11 +125,11 @@ const Vote = ({
         <>{VoteFilledUp}</>
       ) : (
         <BiUpvote
-          title={`upvote ${postMode ? "(a)" : ""}`}
+          title={`${!archived ? `upvote ${postMode ? "(a)" : ""}` : "archived"}`}
           onClick={(e) => castVote(e, true)}
           className={
             (liked ? " text-upvote " : liked === false ? " opacity-50 " : "") +
-            ` flex-none cursor-pointer w-${size} h-${size} hover:text-upvote hover:scale-110 hover:opacity-100`
+            ` flex-none w-${size} h-${size} ${!archived ? "cursor-pointer  hover:text-upvote hover:scale-110 hover:opacity-100" : " opacity-10 "} `
           }
         />
       )}
@@ -151,11 +152,11 @@ const Vote = ({
         <>{VoteFilledDown}</>
       ) : (
         <BiDownvote
-          title={`downvote ${postMode ? "(v)" : ""}`}
+          title={`${!archived ? `downvote ${postMode ? "(v)" : ""}` : `archived`}`}
           onClick={(e) => castVote(e, false)}
           className={
             (liked ? " opacity-50 " : "") +
-            ` flex-none cursor-pointer w-${size} h-${size} hover:text-downvote hover:scale-110 hover:opacity-100`
+            ` flex-none w-${size} h-${size} ${!archived ? "cursor-pointer  hover:text-downvote hover:scale-110 hover:opacity-100" : "opacity-10"} `
           }
         />
       )}